requirements.txt
jupyter
matplotlib
pandas
numpy
timm
torch-lr-finder
tqdm
ultralytics